requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.1.0

# LangChain dependencies
langchain-core>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.10
openai>=1.0.0

# Web scraping and data collection
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0

# Text processing and analysis
textblob>=0.17.1
nltk>=3.8.1
transformers==4.36.2
sentence-transformers==2.2.2

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
wordcloud>=1.9.3

# Vector store and similarity search
faiss-cpu>=1.7.4

# Environment variables
python-dotenv>=1.0.0 